src/adrenalnmodels/requirements.txt
alembic==1.0.5
asn1crypto==0.24.0
blinker==1.4
bcrypt==3.1.7
# cffi==1.15.0
Click==7.0
cryptography
Flask==1.0.2
Flask-Cors
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
paginate==0.5.6
psycopg2-binary
pycparser==2.19
PyJWT==1.4.2
PyMySQL==0.9.2
python-dateutil==2.8.1
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
mistune==2.0.2
boto3~=1.21.34
Authlib~=1.0.0
flask_bcrypt
cachetools==3.1.1
commonmark==0.9.1
deepdiff==5.8.0
dill==0.3.2
dnslib==0.9.19
dnspython==2.2.1
idna==2.10
localstack==0.14.2.1
localstack-client==1.35
localstack-ext==0.14.2.2
ordered-set==4.1.0
plux==1.3.1
psutil==5.9.0
pyaes==1.6.1
pygments==2.11.2
requests==2.25.1
rich==12.2.0
semver==2.13.0
stevedore==3.5.0
tabulate==0.8.9
tailer==0.4.1
typing-extensions==4.2.0
pytz
fpdf==1.7.2
flask-mongoengine
firebase-admin
geographiclib
geopy
xlsxwriter==3.0.3
